+++ src/Retour.php
@@ -88,7 +88,7 @@ class Retour extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '3.0.4';
+    public $schemaVersion = '3.0.6';
 
 
     /**
@@ -524,13 +524,11 @@ class Retour extends Plugin
             'retour/redirects/<siteHandle:{handle}>' => 'retour/redirects/redirects',
 
             'retour/edit-redirect/<redirectId:\d+>' => 'retour/redirects/edit-redirect',
-            'retour/edit-redirect/<redirectId:\d+>/<siteHandle:{handle}>' => 'retour/redirects/edit-redirect',
 
             'retour/add-redirect' => 'retour/redirects/edit-redirect',
-            'retour/add-redirect/<siteHandle:{handle}>' => 'retour/redirects/edit-redirect',
+            'retour/add-redirect/<siteId:\d+>' => 'retour/redirects/edit-redirect',
 
             'retour/delete-redirect/<redirectId:\d+>' => 'retour/redirects/delete-redirect',
-            'retour/delete-redirect/<redirectId:\d+>/<siteHandle:{handle}>' => 'retour/redirects/delete-redirect',
 
             'retour/dashboard' => 'retour/statistics/dashboard',
             'retour/dashboard/<siteHandle:{handle}>' => 'retour/statistics/dashboard',
@@ -552,6 +550,7 @@ class Retour extends Plugin
             '/retour/tables/redirects' => 'retour/tables/redirects',
             // Charts
             '/retour/charts/dashboard/<range:{handle}>' => 'retour/charts/dashboard',
+            '/retour/charts/dashboard/<range:{handle}>/<siteId:\d+>' => 'retour/charts/dashboard',
             '/retour/charts/widget/<days>' => 'retour/charts/widget',
         ];
     }