+++ src/assetbundles/retour/src/vue/DashboardTable.vue
@@ -45,10 +45,17 @@
             'vuetable-pagination-info': VueTablePaginationInfo,
             'vuetable-filter-bar': VueTableFilterBar,
         },
-        props: {},
+        props: {
+            siteId: {
+                type: Number,
+                default: 0,
+            }
+        },
         data: function() {
             return {
-                moreParams: {},
+                moreParams: {
+                    'siteId': this.siteId,
+                },
                 css: {
                     tableClass: 'data fullwidth retour-dashboard',
                     ascendingIcon: 'menubtn retour-menubtn-asc',
@@ -67,16 +74,23 @@
         mounted() {
             this.$events.$on('filter-set', eventData => this.onFilterSet(eventData));
             this.$events.$on('filter-reset', e => this.onFilterReset());
+            // Live refresh the data
+            setInterval(() => {
+                this.$refs.vuetable.refresh();
+            }, 3000);
         },
         methods: {
             onFilterSet (filterText) {
                 this.moreParams = {
-                    'filter': filterText
+                    'siteId': this.siteId,
+                    'filter': filterText,
                 };
                 this.$events.fire('refresh-table', this.$refs.vuetable);
             },
             onFilterReset () {
-                this.moreParams = {};
+                this.moreParams = {
+                    'siteId': this.siteId,
+                };
                 this.$events.fire('refresh-table', this.$refs.vuetable);
             },
             onPaginationData (paginationData) {
@@ -122,7 +136,6 @@
                 return `
                 <a class="add icon" href="${value}" title="Add"></a>
                 `;
-                return '';
             }
         }
     }