+++ src/assetbundles/retour/src/vue/ImportDraggable.vue
@@ -92,6 +92,7 @@
                     'Redirect To',
                     'Match Type',
                     'HTTP Status',
+                    'Site ID',
                 ]
             }
         },