+++ src/migrations/Install.php
@@ -82,7 +82,7 @@ class Install extends Migration
                     'dateUpdated' => $this->dateTime()->notNull(),
                     'uid' => $this->uid(),
 
-                    'locale' => $this->string(12)->defaultValue('en-US'),
+                    'siteId' => $this->integer()->null()->defaultValue(null),
                     'associatedElementId' => $this->integer()->notNull(),
                     'enabled' => $this->boolean()->defaultValue(true),
                     'redirectSrcUrl' => $this->string(255)->defaultValue(''),
@@ -108,7 +108,7 @@ class Install extends Migration
                     'dateUpdated' => $this->dateTime()->notNull(),
                     'uid' => $this->uid(),
 
-                    'locale' => $this->string(12)->defaultValue('en-US'),
+                    'siteId' => $this->integer()->null()->defaultValue(null),
                     'associatedElementId' => $this->integer()->notNull(),
                     'enabled' => $this->boolean()->defaultValue(true),
                     'redirectSrcUrl' => $this->string(255)->defaultValue(''),
@@ -134,6 +134,7 @@ class Install extends Migration
                     'dateUpdated' => $this->dateTime()->notNull(),
                     'uid' => $this->uid(),
 
+                    'siteId' => $this->integer()->null()->defaultValue(null),
                     'redirectSrcUrl' => $this->string(255)->defaultValue(''),
                     'referrerUrl' => $this->string(2000)->defaultValue(''),
                     'remoteIp' => $this->string(45)->defaultValue(''),
@@ -160,21 +161,56 @@ class Install extends Migration
             $this->db->getIndexName(
                 '{{%retour_static_redirects}}',
                 'redirectSrcUrlParsed',
-                true
+                false
             ),
             '{{%retour_static_redirects}}',
             'redirectSrcUrlParsed',
-            true
+            false
         );
+
         $this->createIndex(
             $this->db->getIndexName(
                 '{{%retour_redirects}}',
                 'redirectSrcUrlParsed',
-                true
+                false
             ),
             '{{%retour_redirects}}',
             'redirectSrcUrlParsed',
-            true
+            false
+        );
+
+        $this->createIndex(
+            $this->db->getIndexName(
+                '{{%retour_redirects}}',
+                'siteId',
+                false
+            ),
+            '{{%retour_redirects}}',
+            'siteId',
+            false
+        );
+
+        $this->createIndex(
+            $this->db->getIndexName(
+                '{{%retour_static_redirects}}',
+                'siteId',
+                false
+            ),
+            '{{%retour_static_redirects}}',
+            'siteId',
+            false
+        );
+
+
+        $this->createIndex(
+            $this->db->getIndexName(
+                '{{%retour_stats}}',
+                'siteId',
+                false
+            ),
+            '{{%retour_stats}}',
+            'siteId',
+            false
         );
     }
 
@@ -184,7 +220,7 @@ class Install extends Migration
     protected function addForeignKeys()
     {
         $this->addForeignKey(
-            $this->db->getForeignKeyName('{{%retour_redirects}}', 'siteId'),
+            $this->db->getForeignKeyName('{{%retour_redirects}}', 'associatedElementId'),
             '{{%retour_redirects}}',
             'associatedElementId',
             '{{%elements}}',
@@ -192,10 +228,10 @@ class Install extends Migration
             'CASCADE',
             'CASCADE'
         );
-        /*
+
         $this->addForeignKey(
-            $this->db->getForeignKeyName('{{%retour_redirects}}', 'siteId'),
-            '{{%retour_redirects}}',
+            $this->db->getForeignKeyName('{{%retour_static_redirects}}', 'siteId'),
+            '{{%retour_static_redirects}}',
             'siteId',
             '{{%sites}}',
             'id',
@@ -212,7 +248,6 @@ class Install extends Migration
             'CASCADE',
             'CASCADE'
         );
-        */
     }
 
     /**